refactor: convert usertable to ts
